+++ rnns/rnn_utils.py
@@ -1,5 +1,6 @@
 import torch
 import torch.nn as nn
+from torch.distributions.multivariate_normal import MultivariateNormal
 import matplotlib.pyplot as plt
 import pandas as pd
 import numpy as np
@@ -8,7 +9,6 @@ from copy import deepcopy
 
 # This function preprocesses data to create a training sequences and corresponding
 # DO targets
-
 
 def create_sequence(input_data, train_window):
     seq = []
@@ -20,6 +20,34 @@ def create_sequence(input_data, train_window):
 
     return seq
 
+def build_cov_matrix(var, covs):
+    """
+    This function builds a covariance matrix from variates and covariates
+    """
+    cov_mat = torch.diag(var)
+    cov_mat[0, 1] = 0
+    cov_mat[1, 0] = covs[0]
+    cov_mat[0, 2] = 0
+    cov_mat[2, 0] = covs[1]
+    cov_mat[0, 3] = 0
+    cov_mat[3, 0] = covs[2]
+    cov_mat[1, 2] = 0
+    cov_mat[2, 1] = covs[3]
+    cov_mat[1, 3] = 0
+    cov_mat[3, 1] = covs[4]
+    cov_mat[2, 3] = 0
+    cov_mat[3, 2] = covs[5]
+    # Enforcing matrix to be PD
+    cov_mat = torch.mm(cov_mat, cov_mat.t())
+    cov_mat.add_(torch.eye(len(cov_mat)))
+    try:
+        np.linalg.cholesky(cov_mat.detach().numpy())
+        return cov_mat
+    except:
+        import pdb
+        pdb.set_trace()
+
+
 
 class LSTM(nn.Module):
     def __init__(self, input_size=1, hidden_layer_size=500, fc_size=500, output_size=1):
@@ -30,9 +58,7 @@ class LSTM(nn.Module):
 
         self.fc_0 = nn.Linear(hidden_layer_size, fc_size)
         self.fc_1 = nn.Linear(fc_size, fc_size)
-        self.fc_2 = nn.Linear(fc_size, fc_size)
-        self.fc_3 = nn.Linear(fc_size, fc_size)
-        self.fc_4 = nn.Linear(fc_size, output_size)
+        self.fc_2 = nn.Linear(fc_size, output_size)
 
         self.hidden_cell = (torch.zeros(1, 1, self.hidden_layer_size),
                             torch.zeros(1, 1, self.hidden_layer_size))
@@ -46,15 +72,17 @@ class LSTM(nn.Module):
         out = self.fc_0(out)
         out = self.relu(out)
         out = self.dropout(out)
-        #out = self.fc_1(out)
-        #out = self.relu(out)
-        #out = self.dropout(out)
-        #out = self.fc_2(out)
-        #out = self.relu(out)
-        #out = self.dropout(out)
-        #out = self.fc_3(out)
-        #out = self.relu(out)
-        #out = self.dropout(out)
-        out = self.fc_4(out.view(len(input_seq), -1))
+        out = self.fc_1(out)
+        out = self.relu(out)
+        out = self.dropout(out)
+        out = self.fc_2(out.view(len(input_seq), -1))
         out = self.dropout(out)
         return out[-1]
+
+def build_dist(model, seq):
+    y_pred = model(seq).view(-1)
+    mu = y_pred[:4]
+    var = torch.abs(y_pred[4:8])
+    covs = y_pred[-6:]
+    cov_matrix = build_cov_matrix(var, covs)
+    return MultivariateNormal(mu, cov_matrix)
